dashes

+++ apps/ios/SimpleXChat/APITypes.swift
@@ -2335,33 +2335,33 @@ public struct SMPServerSubs: Codable {
 }
 
 public struct AgentSMPServerStatsData: Codable {
-    public var _sentDirect: Int64
-    public var _sentViaProxy: Int64
-    public var _sentProxied: Int64
-    public var _sentDirectAttempts: Int64
-    public var _sentViaProxyAttempts: Int64
-    public var _sentProxiedAttempts: Int64
-    public var _sentAuthErrs: Int64
-    public var _sentQuotaErrs: Int64
-    public var _sentExpiredErrs: Int64
-    public var _sentOtherErrs: Int64
-    public var _recvMsgs: Int64
-    public var _recvDuplicates: Int64
-    public var _recvCryptoErrs: Int64
-    public var _recvErrs: Int64
-    public var _ackMsgs: Int64
-    public var _ackAttempts: Int64
-    public var _ackNoMsgErrs: Int64
-    public var _ackOtherErrs: Int64
-    public var _connCreated: Int64
-    public var _connSecured: Int64
-    public var _connCompleted: Int64
-    public var _connDeleted: Int64
-    public var _connDelAttempts: Int64
-    public var _connDelErrs: Int64
-    public var _connSubscribed: Int64
-    public var _connSubAttempts: Int64
-    public var _connSubErrs: Int64
+    public var _sentDirect: Int
+    public var _sentViaProxy: Int
+    public var _sentProxied: Int
+    public var _sentDirectAttempts: Int
+    public var _sentViaProxyAttempts: Int
+    public var _sentProxiedAttempts: Int
+    public var _sentAuthErrs: Int
+    public var _sentQuotaErrs: Int
+    public var _sentExpiredErrs: Int
+    public var _sentOtherErrs: Int
+    public var _recvMsgs: Int
+    public var _recvDuplicates: Int
+    public var _recvCryptoErrs: Int
+    public var _recvErrs: Int
+    public var _ackMsgs: Int
+    public var _ackAttempts: Int
+    public var _ackNoMsgErrs: Int
+    public var _ackOtherErrs: Int
+    public var _connCreated: Int
+    public var _connSecured: Int
+    public var _connCompleted: Int
+    public var _connDeleted: Int
+    public var _connDelAttempts: Int
+    public var _connDelErrs: Int
+    public var _connSubscribed: Int
+    public var _connSubAttempts: Int
+    public var _connSubErrs: Int
 }
 
 public struct XFTPServersSummary: Codable {
@@ -2388,16 +2388,16 @@ public struct XFTPServerSummary: Codable, Identifiable {
 }
 
 public struct AgentXFTPServerStatsData: Codable {
-    public var _uploads: Int64
+    public var _uploads: Int
     public var _uploadsSize: Int64
-    public var _uploadAttempts: Int64
-    public var _uploadErrs: Int64
-    public var _downloads: Int64
+    public var _uploadAttempts: Int
+    public var _uploadErrs: Int
+    public var _downloads: Int
     public var _downloadsSize: Int64
-    public var _downloadAttempts: Int64
-    public var _downloadAuthErrs: Int64
-    public var _downloadErrs: Int64
-    public var _deletions: Int64
-    public var _deleteAttempts: Int64
-    public var _deleteErrs: Int64
+    public var _downloadAttempts: Int
+    public var _downloadAuthErrs: Int
+    public var _downloadErrs: Int
+    public var _deletions: Int
+    public var _deleteAttempts: Int
+    public var _deleteErrs: Int
 }
